requirements_dev.txt
pytest==3.3.1
pytest-flask==0.10.0
flake8==3.5.0
flake8-import-order==0.16
invoke==0.22.0
requests==2.18.4
